conversion to datetime seems to work
